scrapers/gdelt_scrapers/requirements.txt
ago==0.0.93
attrs==21.4.0
Automat==20.2.0
awscli==1.22.96
beautifulsoup4==4.11.1
botocore==1.24.41
bs4==0.0.1
cachetools==5.0.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==36.0.2
cssselect==1.1.0
docutils==0.15.2
dotmap==1.3.30
elastic-transport==8.1.1
elasticsearch==8.1.2
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
hjson==3.0.2
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.5.0
itemloaders==1.0.4
jieba3k==0.35.1
jmespath==1.0.0
joblib==1.1.0
langdetect==1.0.9
lxml==4.8.0
news-please==1.5.22
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.3
packaging==21.3
parsel==1.6.0
Pillow==9.1.0
plac==1.3.5
Protego==0.2.1
proto-plus==1.20.3
protobuf==3.20.0
psycopg2-binary==2.9.3
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
pywin32==303
PyYAML==5.4.1
queuelib==1.6.2
readability-lxml==0.8.1
regex==2022.3.15
requests==2.27.1
requests-file==1.5.1
rsa==4.7.2
s3transfer==0.5.2
Scrapy==2.6.1
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tinysegmenter==0.3
tldextract==3.2.1
tqdm==4.64.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.1.1
urllib3==1.26.9
w3lib==1.22.0
warcio==1.7.4
zope.interface==5.4.0
